DBProblemsDemoPython/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.3
mysql-connector-python==8.0.26
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-instrumentation-mysql==0.41b0
opentelemetry-instrumentation-requests==0.41b0
requests==2.28.2
protobuf>=3.20.0